requirements.txt
python-dotenv
langchain
openai
google-api-python-client # Google API
google-search-results # SerpAPI
youtube-transcript-api
chromadb
tiktoken
faiss-cpu
wikipedia
#psycopg2
psycopg2-binary
#pinecone-client
pgvector
google-generativeai
langchain_community
langchain-google-genai
